+++ exchangerate/internal/wrapper/wrapper.go
@@ -2,9 +2,11 @@ package wrapper
 
 type result struct {
 	Err   error
-	Value float64
+	Value float32
 }
 
+// Wraps external services that we can query for exchange rate.
+// returns a readonly channel where the result of the retrieve can get passed.
 type Wrapper interface {
 	GetExchangeRate(pair string) <-chan result
 }
